backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
supabase==1.1.1
yt-dlp==2023.7.6
pydub==0.25.1
python-dotenv==1.0.0